Make it work with Julia v1.0.

I get
```
 ─────────────────────────────────────────────────────────────────────────────
                                                    Time         Allocations
                                               ──────────────   ──────────────
                Total measured:                   8169219s         21.4GiB

 Section                               ncalls     time   %tot     alloc   %tot
 ─────────────────────────────────────────────────────────────────────────────
 run simulation 1                           1    80.2s   100%   19.1GiB  100%
   assemble                                 1    39.6s  49.4%   9.07GiB  47.5%
     assemble problems                      1    32.3s  40.3%   3.50GiB  18.3%
     construct global assemblies            1    7.31s  9.12%   5.56GiB  29.1%
       get_field_assembly                   1    5.88s  7.33%   4.14GiB  21.7%
       sum K                                1    1.19s  1.48%   1.32GiB  6.89%
       get_boundary_assembly                1    236ms  0.29%   94.6MiB  0.48%
       sum f                                1   2.75ms  0.00%   13.4MiB  0.07%
   solution                                 1    23.2s  28.9%   7.64GiB  40.0%
     factorize K                            1    18.5s  23.0%   5.84GiB  30.6%
     update solution                        1    1.25s  1.55%    255MiB  1.31%
     create symmetric K                     1    1.24s  1.55%   1.32GiB  6.90%
     eliminate boundary conditions ...      1    830ms  1.04%    125MiB  0.64%
     solve u                                1    627ms  0.78%   56.1MiB  0.29%
     solve la                               1    340ms  0.42%   53.5MiB  0.27%
   initialize model                         1    7.98s  9.95%   1.54GiB  8.05%
     parse input data                       1    5.32s  6.63%   1.03GiB  5.37%
     initialize models                      1    2.66s  3.32%    524MiB  2.68%
   write results to disk                    1    5.51s  6.87%    774MiB  3.96%
 ─────────────────────────────────────────────────────────────────────────────
```